requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.4
orjson==3.8.3
pydantic==1.10.2
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
yarl==1.8.2
